requirements.txt
urllib3
python-taiga==1.3.3
requests
python-dotenv
markdown
